LinkedQueue

+++ ArrayQueue.cpp
@@ -127,7 +127,7 @@ void ArrayQueue::imprime(){// metodo para imprimir los elementos de la cola
 
 void ArrayQueue::anula()//metodo para borrar los elementos de la cola
 {
-    for (size_t i = 0; i < capacidad; i++){//elimina elementos de la cola
+    for (int i = 0; i < capacidad; i++){//elimina elementos de la cola
         delete cola[i];
     }
